requirements.txt
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
requests>=2.31
pydantic>=2.7
